third_party/blink/web_tests/http/tests/devtools/sources/debugger-ui/show-function-definition-expected.txt

Tests that "Show Function Definition" jumps to the correct location.


Running: testRevealFunctionDefinition
Function location revealed: [15:24]

Running: testDumpFunctionDefinition
jumpToMe
ƒ jumpToMe()
      {
          var result = 12345;
          return window.foo || result;
      }
